fix: improve `format` parsing
